sample soundActs.txt file
This is the config fle 3sound_pyg and 4sound_pyg use to set up the 2 acts and set the number of buttons on each screen as well state the text for the buttons and which audio files to play.  I was originally using Pygame for the audio and read it likes .ogg files. But now using omxplayer so the option for audio file formats is a lot broader.

As a side note I used Audacity on the Pi 3B+ to do some of the conversion and minor editing of the audio files.